requirements.txt
playwright
beautifulsoup4
pymongo
python-dotenv
pydantic
PyMuPDF
openai